Stop tangling literate config from init.el
Previously my config consisted of three main files:

1. `early-init.el` which disabled `package.el`.
2. `init.el` which loaded `straight.el`, `use-package` and `org` to
then tangle from `config.org`.
3. `config.el` which was the tangled config.

There were several consequences to this system. In particular, we
needed to load org BEFORE calling `org-babel-load-file`. In addition
it is confusing to following the load order of packages split across
three files. Finally, this repository and `config.org` had to be in
the emacs config directory for `org-babel-load-file` to tangle and
source the file properly.

Now you have to manually tangle the config but everything should be
much clearer with better separation between the system/live config and
the version-controlled source. Since org-mode is built into emacs now,
it should always be available after first install to allow the initial
tangling.