Remove elpaca.el from config
Emacs 28 on MacOS Ventura has a bug related to SVG support. With
this bug I have to evaluate the generated `init.el` manually after
first startup so I can't use an alternative package manager that
relies on disabling `package.el` in the `early-init.el`. This
commit removes `elpaca.el` and uses the built-in `package.el` for
this reason. `asdf.el` and `copilot.el` must be cloned and loaded
locally using this setup.