Add pipx and the jedi-language-server
Add pipx which is akin to brew or npx but for python packages to Linux
box. Install the jedi language server with it. This is in preparation
for Advent of Code 2023. I chose the jedi language server because it is
backed by the jedi project which is widely used, well-backed and fully
open-source (unlike pyright which is tied to microsoft). Also as the
authors claim, it doesn't depend on a lot of 3rd party modules like the
python-language-server so there are fewer issues with various
integrations and broken tooling.