Gnus email configuration
Rather than using the `isync` + `notmuch` + `msmtp` combo, this commit
sets up Gnus inside Emacs for email support. In particular just for two
of my simplest IMAP emails.

This largely reverts the efforts of my previous commit. However I still
rely on Age for encryption/decryption.

The reasoning behind using Gnus is that it is self-contained by Emacs
and should be more easily portable to other platforms; especially given
I have very simple email requirements. In addition, it can be
configured to also serve as my RSS reader.

The Gnus configuration contains somewhat sensitive information so it is
encrypted by Chezmoi. It also relies on `.authinfo.age` for
authentication information. That file is encrypted both in this
repository and on my local filesystem to avoid keeping passwords in
plain text.