Switch to transient over repeat mode or embark
Transient gives a much finer-grained control over sticky menus than
repeat or embark. Repeat mode is nice because it is already enabled for
some common commands and it is integrated with `use-package` so creating
repeat maps is very simple. However repeat mode is really designed to
work for a single package and still requires you to know at least the
entrypoint to that command. Embark could replace and has the potential
to be the nicest option but has no convenient way to handle keybinding
collison between different detected maps that may appear when you've
configured it heavily.

Transient is built into emacs and can be used to achieve similar
results; albeit with more complexity in the configuration. In particular
you can define transient maps to which you can append suffixes from
other packages. These maps can be entered directly via keybind or you
can advice entrypoint functions to invoke the map. Since you're advising
the function (instead of the keybinding), even if you invoke the
function from `M-x`, you'll still get access to the map.