Use Rose Pine themes
Use [Rose Pine themes](https://github.com/rose-pine/rose-pine-theme)
in terminal emulator and Helix. Ghostty can
automatically shift between Rose Pine and Rose Pine
Dawn (dark and light themes) based on time of day.
Helix uses Rose Pine Moon which is a muted dark
theme. The idea is that it is immediately visually
obvious when I am in Helix. The fact that it's not
a proper light theme during the day is only ok if I
truly rely on Claude Code for most edits and pop
into Helix only to adjust.